fix(cli): close 5 clippy errors surfaced by 1.88 on CI
CI rustc 1.88 clippy is stricter than local (1.95 emits a superset but
some lints are 1.88-only, or differently prioritized). Five errors on
crates/1bit-cli:

- install.rs:44   dead_code fields size_mb, license → #[allow(dead_code)]
                  with comment explaining round-trip fidelity with
                  packages.toml
- install_model.rs:37  dead_code fields description, requires →
                  struct-level #[allow(dead_code)] + mirroring note
- install.rs:158  collapsible_if → let-chain
- power.rs:227    collapsible_if → let-chain
- update.rs:385   collapsible_if → let-chain
- preflight.rs:132  unnecessary_cast u64 → u64 (libc statvfs fields
                  are already u64 on Linux, casts were no-ops)

cargo clippy -p onebit-cli --all-targets -- -D warnings now clean.

+++ crates/1bit-cli/src/preflight.rs
@@ -129,7 +129,9 @@ impl SystemProbe for RealProbe {
         }
         // f_bavail = blocks available to non-root; f_frsize = fragment
         // size in bytes. Multiply to bytes, then /1 GiB.
-        let bytes = (st.f_bavail as u64).saturating_mul(st.f_frsize as u64);
+        // (libc already types these as u64 on Linux — explicit casts are
+        // a no-op and clippy flags them.)
+        let bytes = st.f_bavail.saturating_mul(st.f_frsize);
         bytes / (1024 * 1024 * 1024)
     }
